ARM: OMAP2+: pm33xx: Add base cpuidle support
In some cases, like that of C1 and greater c-states on am335x, we must
use the wkup-m3 and sleep33xx code to properly get the cpu to low power
state. In order to do this we place update the pm33xx code to support
calling into the sleep33xx code without using the additional supoprt
code necessary for suspend and also and a new op to be defined in
pm33xx-core, cpu_suspend, which will allow us to make a call to just
suspend the cpu without of the support code required for soc suspend.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>